refactor(api): move view mappers onto response-record from() factories
SailOperations held ~9 pure static view-mappers that each built one response
DTO from a domain object. Relocate them onto the records themselves as
from() factories — AgentConfigView, AgentStatusResponse, AgentRunView,
AgentReportResponse, SpecView, DispatchedSpecView, AgentStatusView,
SpecSummaryView, BoardSummaryView — matching the eight from() factories the
records already carry (RunView, ReviewView, GlobalSpecView, ...).

Each DTO now owns its construction from the domain (SRP); SailOperations
keeps only orchestration and shrinks from 1511 to 1404 lines. Pure move, no
behavior change: the endpoint tests that covered the mappers as privates now
cover them as factories, and mvn clean verify is green with all coverage
checks met (the factories land in the strictly-100% ApiModels).
